Refactored portals and updated main scene
Refactored portals, now they are saved into separate scenes and are much easier to use.
Also updated main scene to be more clear in what happening in environment and portals.
Reconfigured lightning and environment on camera, now it finally looks to be working as intended.
Added FPS counter to track performance issues when looking at two portals. Will need to investigate it later and fix if possible, of course.