[Godot] Updated visuals to Minecraft-like style
Updated visuals in demo scene to Minecraft-like style, with blocks used for objects models and textures directly from Minecraft (I suppose).
Also, noticed one another bug earlier and will write it here now in order not to forget. When any object is thrown through portal, for a moment it can be seen in a plane of connected portal, which should not happen, as connected portal should stream the backside of current portal, from where this object should instantly disappear, when passing portal plane. So it is pretty strange and weird thing and I will need to investigate it later also, in addition to all of the other previous and future bugs ang problems, yeah, huh...