[Godot] Updated map, fixed portals and started implementing water
Updated map with some coordinates fixes and new pool for water simulation testing.
Fixed portals glitches, looks like the problem was in CSGBox3D width, and it should be always set to 0.
Started implementing water, only a simple plane with shader for now, but it is already not working if importing it as a scene instance, ugh... Will need to investigate and fix this later, or development will be much harder and awkward, I suppose.